move utilites/utilites.hpp to legacy
