+++ cpp/src/io/soa/point_soa_reader.cu
@@ -23,7 +23,7 @@
 #include <cudf/utilities/error.hpp>
 #include <rmm/rmm.h>
 #include <cuspatial/soa_readers.hpp>
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 
 namespace cuspatial
 {
