+++ cpp/src/io/soa/timestamp_soa_reader.cu
@@ -7,7 +7,7 @@
 #include <cudf/types.h>
 #include <cudf/legacy/column.hpp>
 #include <cuspatial/soa_readers.hpp>
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 
 namespace cuspatial
 {
