+++ cpp/src/spatial/legacy/haversine.cu
@@ -23,7 +23,7 @@
 #include <utilities/legacy/cuda_utils.hpp>
 #include <type_traits>
 #include <thrust/device_vector.h>
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 #include <cuspatial/legacy/haversine.hpp>
 #include <cuspatial/error.hpp>
