+++ cpp/src/spatial/legacy/point_in_polygon.cu
@@ -17,7 +17,7 @@
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
 #include <utilities/legacy/cuda_utils.hpp>
 #include <type_traits>
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 #include <cuspatial/legacy/point_in_polygon.hpp>
 #include <cuspatial/error.hpp>
