+++ cpp/src/trajectory/legacy/derive_trajectories.cu
@@ -20,7 +20,7 @@
 #include <utilities/legacy/cuda_utils.hpp>
 #include <rmm/thrust_rmm_allocator.h>
 
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 #include <utility/trajectory_thrust.cuh>
 #include <cuspatial/legacy/trajectory.hpp>
 #include <cuspatial/error.hpp>
