+++ cpp/src/utility/legacy/utility.hpp
