+++ cpp/tests/spatial/legacy/hausdorff_test.cu
@@ -22,7 +22,7 @@
 
 #include <cuspatial/soa_readers.hpp> 
 #include <cuspatial/legacy/hausdorff.hpp>
-#include <utility/utility.hpp> 
+#include <utility/legacy/utility.hpp>
 #include "hausdorff_util.h" 
 
 #include <tests/utilities/legacy/cudf_test_utils.cuh>
